requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
openai
twilio==8.10.0
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0
